+++ geotrek/core/serializers.py
@@ -32,6 +32,7 @@ class PathGeojsonSerializer(MapentityGeojsonModelSerializer):
 class TrailSerializer(DynamicFieldsMixin, serializers.ModelSerializer):
     length = serializers.FloatField(source='length_display')
     name = serializers.CharField(source='name_display')
+    category = serializers.SlugRelatedField('label', read_only=True)
 
     class Meta:
         model = Trail